Alterando IDs de int para long

+++ FiapCloudGamesAPI/Entidades/Dtos/UsuarioDto.cs
@@ -9,7 +9,7 @@ namespace FiapCloudGamesAPI.Entidades.Dtos
         public string Apelido { get; set; }
         public string Email { get; set; }
         public DateTime DataNascimento { get; set; }
-        public int PerfilId { get; set; }
+        public long PerfilId { get; set; }
         public Perfil Perfil { get; set; }
         public ICollection<Avaliacao> Avaliacoes { get; set; }
         public ICollection<BibliotecaDoJogador> Biblioteca { get; set; } // Relacionamento 1:N
